change Toast to english add alerts

+++ app/src/main/java/com/example/projektzd/fragments/databaseFragments/EntityFragment.kt
@@ -1,7 +1,11 @@
 package com.example.projektzd.fragments.databaseFragments
 
 import android.annotation.SuppressLint
+import android.app.AlertDialog
 import android.app.DatePickerDialog
+import android.app.Dialog
+import android.content.DialogInterface
+import android.os.Build
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -9,6 +13,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.Toast
+import androidx.annotation.RequiresApi
 import androidx.core.net.toUri
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
@@ -25,6 +30,7 @@ import java.lang.Exception
 import java.text.SimpleDateFormat
 import java.time.LocalDate
 import java.time.LocalDateTime
+import java.time.ZoneId
 import java.time.format.DateTimeFormatter
 import java.time.temporal.ChronoUnit
 import java.util.*
@@ -54,6 +60,8 @@ class EntityFragment(
         binding.pageCount.text = book.numberOfPages.toString()
         binding.rentalDateTxt.text = book.rentalDate
         binding.returnDateTxt.text = book.returnDate
+        binding.author.text = book.authors
+        binding.description.text = book.description
         rentalDateString = book.rentalDate
         returnDateString = book.returnDate
         changeFav(binding.fillHeart)
@@ -143,8 +151,22 @@ class EntityFragment(
                     dbHelper.updateRead(book.id, 1)
                     imageView.setImageResource(R.drawable.read_icon)
                 } else {
-                    dbHelper.updateRead(book.id, 0)
-                    imageView.setImageResource(R.drawable.unread_icon)
+                    var builder = AlertDialog.Builder(activity)
+                    builder.setTitle("Do you want to read this book again")
+                    builder.setMessage(book.title)
+                    //builder.setCancelable(true)
+                    builder.setPositiveButton("YES"){dialog, which->
+                        dbHelper.updateRead(book.id, 0)
+                        imageView.setImageResource(R.drawable.unread_icon)
+                        Toast.makeText(activity,"${book.title} is unread",Toast.LENGTH_SHORT).show()
+                    }
+                    builder.setNegativeButton("No"){dialog,which ->
+                        Toast.makeText(activity,"nothing has changed",Toast.LENGTH_SHORT).show()
+                    }
+                    val dialog: AlertDialog = builder.create()
+
+                    dialog.show()
+
                 }
 
             }
@@ -157,9 +179,21 @@ class EntityFragment(
 
     fun handleDelete() {
         binding.deleteBtn.setOnClickListener {
-            dbHelper.deleteData(book.id)
-            supportFragmentManager.popBackStack()
-            showToast("Książka została usunięta")
+            var builder = AlertDialog.Builder(activity)
+            builder.setTitle("Do you want to delete this book")
+            builder.setMessage(book.title)
+            builder.setPositiveButton("YES"){dialog, which->
+                dbHelper.deleteData(book.id)
+                supportFragmentManager.popBackStack()
+                showToast("${book.title} was removed")
+            }
+            builder.setNegativeButton("No"){dialog,which ->
+                Toast.makeText(activity,"nothing has changed",Toast.LENGTH_SHORT).show()
+            }
+            val dialog: AlertDialog = builder.create()
+
+            dialog.show()
+
         }
     }
 
@@ -198,7 +232,7 @@ class EntityFragment(
     ) {
         binding.rentalDateBtn.setOnClickListener {
             activity?.let { it1 ->
-                DatePickerDialog(
+               var dpd = DatePickerDialog(
                         it1,
                         DatePickerDialog.OnDateSetListener { view, mYear, mMonth, mDay ->
                             c.set(Calendar.YEAR, mYear)
@@ -212,10 +246,13 @@ class EntityFragment(
                         month,
                         day
                 )
-            }?.show()
+                dpd.datePicker.maxDate=System.currentTimeMillis()
+                dpd.show()
+            }
         }
     }
 
+    @RequiresApi(Build.VERSION_CODES.O)
     private fun handleReturnDate(
             c: Calendar,
             year: Int,
@@ -225,7 +262,7 @@ class EntityFragment(
     ) {
         binding.returnDateBtn.setOnClickListener {
             activity?.let { it1 ->
-                DatePickerDialog(
+              var dpd =  DatePickerDialog(
                         it1,
                         DatePickerDialog.OnDateSetListener { view, mYear, mMonth, mDay ->
                             c.set(Calendar.YEAR, mYear)
@@ -239,7 +276,11 @@ class EntityFragment(
                         month,
                         day
                 )
-            }?.show()
+
+                dpd.datePicker.minDate=System.currentTimeMillis()
+                dpd.show()
+            }
+
         }
     }
